Move common functionality into bast AuthAction trait Add constraint, AuthAction needs to be mixed into a Controller
